+++ R/NetLogoR-package.R
@@ -1,5 +1,7 @@
 #' The \code{NetLogoR} package
 #'
+#' The suggested package \pkg{fastshp} can be installed with
+#' \code{install.packages("fastshp", repos = "http://rforge.net", type = "source")}.
 "_PACKAGE"
 
 #' @import methods