+++ R/turtle-functions.R
@@ -3182,3 +3182,4 @@ setMethod(
                                    [3:ncol(turtles@.Data)])
     return(spdf)
   })
+