+++ NEWS.md
@@ -2,7 +2,11 @@ Known issues: https://github.com/PredictiveEcology/NetLogoR/issues
 
 version 0.3.0
 =============
+* add quickPlot version >= 0.1.1.9000 dependency
 * update dependencies on `SpaDES`-related packages to only include the ones actually used.
+* Define [[ and $ for `worldArray` to extract subset layers.
+* Add `show` methods for `worldMatrix` and `worldArray`, similar to `RasterLayer` and `RasterStack`
+* Add tools so `quickPlot::Plot` works, e.g., `Plot(agentMatrixObj)` or `Plot(worldMatrixObj)` or `Plot(worldArrayObj)`, `Plot(worldArrayObj$layer1)`. `addTo` argument implemented for layering, `Plot(worldMatrixObj); Plot(agentMatrixObj, addTo="worldMatrixObj")` will plot the agents on the map.
 
 version 0.2.0
 =============